Added and began using Point and Vector classes instead of individual x,y,z components, moved almost all drone functionality into Model_Drone.

+++ followApp/src/main/java/edu/illinois/mitra/demo/follow/TestDrawer.java
@@ -9,7 +9,7 @@ import edu.illinois.mitra.starl.interfaces.LogicThread;
 import edu.illinois.mitra.starl.objects.ItemPosition;
 import edu.illinois.mitra.starl.objects.ObstacleList;
 import edu.illinois.mitra.starl.objects.Obstacles;
-import edu.illinois.mitra.starl.objects.Point3d;
+import edu.illinois.mitra.starl.objects.Point3i;
 import edu.illinois.mitra.starlSim.draw.Drawer;
 
 public class TestDrawer extends Drawer {
@@ -52,8 +52,8 @@ public class TestDrawer extends Drawer {
 			else
 				g.setColor(Color.GRAY);
 
-			Point3d nextpoint = currobs.obstacle.firstElement();
-			Point3d curpoint ;
+			Point3i nextpoint = currobs.obstacle.firstElement();
+			Point3i curpoint ;
 			int[] xs = new int[currobs.obstacle.size()];
 			int[] ys = new int[currobs.obstacle.size()];
