+++ raceApp/src/main/java/edu/illinois/mitra/demo/race/RaceDrawer.java
@@ -10,7 +10,7 @@ import edu.illinois.mitra.starl.interfaces.LogicThread;
 import edu.illinois.mitra.starl.objects.ItemPosition;
 import edu.illinois.mitra.starl.objects.ObstacleList;
 import edu.illinois.mitra.starl.objects.Obstacles;
-import edu.illinois.mitra.starl.objects.Point3d;
+import edu.illinois.mitra.starl.objects.Point3i;
 import edu.illinois.mitra.starlSim.draw.Drawer;
 
 public class RaceDrawer extends Drawer {
@@ -53,8 +53,8 @@ public class RaceDrawer extends Drawer {
 			else
 				g.setColor(Color.GRAY);
 			
-			Point3d nextpoint = currobs.obstacle.firstElement();
-			Point3d curpoint = currobs.obstacle.firstElement();
+			Point3i nextpoint = currobs.obstacle.firstElement();
+			Point3i curpoint = currobs.obstacle.firstElement();
 			int[] xs = new int[currobs.obstacle.size()]; 
 			int[] ys = new int[currobs.obstacle.size()];
